requirements.txt
streamlit
requests
pandas
plotly
openai
langchain
langgraph
langchain-openai
langgraph
python-dotenv
langchain-tavily